+++ src/components/iconButton/index.js
@@ -1,0 +1,4 @@
+import IconButton from './IconButton';
+
+export default IconButton;
+export { IconButton };
