+++ src/components/iconButton/theme.css
@@ -1,0 +1,27 @@
+@import '@teamleader/ui-colors';
+
+.icon-button {
+  background-color: transparent;
+  border: 0;
+
+  &.is-disabled {
+    cursor: auto;
+    opacity: 0.48;
+    pointer-events: none;
+  }
+
+  &:not(.is-disabled) {
+    &:hover {
+      background-color: color(var(--color-neutral-darkest) a(18%));
+    }
+
+    &:focus {
+      box-shadow: 0 0 0 2px color(var(--color-neutral-darkest) a(24%));
+    }
+
+    &:active {
+      box-shadow: inset 0 2px 3px color(var(--color-neutral-darkest) a(12%));
+      background-color: color(var(--color-neutral-darkest) a(18%));
+    }
+  }
+}
