+++ src/components/qTip/QTip.js
@@ -2,7 +2,8 @@ import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
 import theme from './theme.css';
-import Button, { IconButton } from '../button';
+import Button from '../button';
+import IconButton from '../iconButton';
 import Overlay from '../overlay';
 import { IconArrowRightSmallOutline } from '@teamleader/ui-icons';
 import Box from '../box';
