+++ src/components/toast/Toast.js
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { IconButton } from '../button';
+import IconButton from '../iconButton';
 import Link from '../link';
 import { TextBody } from '../typography';
 import LoadingSpinner from '../loadingSpinner';
